+++ .gitignore
@@ -3,3 +3,4 @@ prentisapp-50e4e-firebase-adminsdk-vvo4f-67dfbe422e.json
 *.xcuserstate
 server/config.js
 server/node_modules/
+*.un~
