+++ Prentis/Prentis/Controllers/HomePageViewController.swift
@@ -12,6 +12,8 @@ import FirebaseAuth
 import FirebaseFirestore
 import FirebaseStorage
 import FaceAware
+import Alamofire
+import PusherSwift
 
 class HomePageViewController: UIViewController, UITableViewDataSource, UITableViewDelegate {
 
@@ -21,9 +23,12 @@ class HomePageViewController: UIViewController, UITableViewDataSource, UITableVi
     var db = Firestore.firestore()
     var documents = [] as [[String: Any]]
     var caller: Mentor!
+    var pusher : Pusher!
+
     
     override func viewDidLoad() {
         super.viewDidLoad()
+        listenForCall()
         self.navigationController?.navigationBar.layer.masksToBounds = false
         self.navigationController?.navigationBar.layer.shadowColor = UIColor.black.cgColor
         self.navigationController?.navigationBar.layer.shadowOpacity = 0.8
@@ -58,37 +63,37 @@ class HomePageViewController: UIViewController, UITableViewDataSource, UITableVi
         userTable.dataSource = self
         userTable.delegate = self
         let mentorUID = Auth.auth().currentUser!.uid
-        db.collection("User").document(mentorUID).collection("Call").whereField("status", isEqualTo: "calling").addSnapshotListener { QuerySnapshot, error in
-            print("hi did you run")
-            guard let documents = QuerySnapshot?.documents else {
-                print("Error fetching documents: \(error!)")
-                return
-            }
-            if documents.count > 0 {
-                let channelName = documents[0].documentID
-                let menteeUID = documents[0].data()["menteeUID"] as! String
-                //self.db.collection("User").document(Auth.auth().currentUser!.uid).collection("Call").document(channelName).updateData(["status": "accepted"])
-                
-                let docRef = self.db.collection("User").document(menteeUID)
-                
-                docRef.getDocument { (document, error) in
-                    if let document = document, document.exists {
-                        let dataDescription = document.data().map(String.init(describing:)) ?? "nil"
-                        self.caller = Mentor(mentorDoc: document.data() as! [String: Any])
-                        self.performSegue(withIdentifier: "receiveSegue", sender: self)
-                        print("Document data: \(dataDescription)")
-                    } else {
-                        print("Document does not exist")
-                    }
-                }
-                
-            } else {
-                print("NOTHING TO DO HERE")
-            }
-            
-            
-            
-        }
+//        db.collection("User").document(mentorUID).collection("Call").whereField("status", isEqualTo: "calling").addSnapshotListener { QuerySnapshot, error in
+//            print("hi did you run")
+//            guard let documents = QuerySnapshot?.documents else {
+//                print("Error fetching documents: \(error!)")
+//                return
+//            }
+//            if documents.count > 0 {
+//                let channelName = documents[0].documentID
+//                let menteeUID = documents[0].data()["menteeUID"] as! String
+//                //self.db.collection("User").document(Auth.auth().currentUser!.uid).collection("Call").document(channelName).updateData(["status": "accepted"])
+//
+//                let docRef = self.db.collection("User").document(menteeUID)
+//
+//                docRef.getDocument { (document, error) in
+//                    if let document = document, document.exists {
+//                        let dataDescription = document.data().map(String.init(describing:)) ?? "nil"
+//                        self.caller = Mentor(mentorDoc: document.data() as! [String: Any])
+//                        self.performSegue(withIdentifier: "receiveSegue", sender: self)
+//                        print("Document data: \(dataDescription)")
+//                    } else {
+//                        print("Document does not exist")
+//                    }
+//                }
+//
+//            } else {
+//                print("NOTHING TO DO HERE")
+//            }
+//
+//
+//
+//        }
         
     }
     
@@ -179,6 +184,43 @@ class HomePageViewController: UIViewController, UITableViewDataSource, UITableVi
         
         // Get the new view controller using segue.destination.
         // Pass the selected object to the new view controller.
+    }
+    
+    private func listenForCall() {
+        let key = "200b87f4cd87ab883b75"
+        let cluster = "us2"
+        let uid = Auth.auth().currentUser!.uid
+        print("this whole method was called")
+        print("UID: \(uid)")
+        pusher = Pusher(key: key, options: PusherClientOptions(host: .cluster(cluster)))
+        
+        let channel = pusher.subscribe("calls")
+        let _ = channel.bind(eventName: uid, callback: { (data: Any?) -> Void in
+            if let data = data as? [String: AnyObject] {
+                
+                print("hey I just got some shit!!!!!!!!!!!!!!!!")
+                
+                let channelName = data["channel"] as! String
+                let menteeUID = data["caller"] as! String
+                
+                let docRef = self.db.collection("User").document(menteeUID)
+                
+                docRef.getDocument { (document, error) in
+                    if let document = document, document.exists {
+                        let dataDescription = document.data().map(String.init(describing:)) ?? "nil"
+                        self.caller = Mentor(mentorDoc: document.data() as! [String: Any])
+                        self.performSegue(withIdentifier: "receiveSegue", sender: self)
+                        print("Document data: \(dataDescription)")
+                    } else {
+                        print("Document does not exist")
+                    }
+                }
+                
+                
+            }
+        })
+        
+        pusher.connect()
     }
 
 }
